feat: added a heartbeat
Crystal sockets weren't being closed correctly and were stuck in CLOSED_WAIT state.
Adding a heartbeat will now ping the connected on an interval and clean up the sockets
if they are no longer active. Some of the close commands we had in the code were no
longer needed so those got cleaned up as well.